pyproject.toml
[project]
name = "rpi-sms-receiver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[all]>=0.115.11",
    "pyserial>=3.5",
    "python-dotenv>=1.0.1",
]
